feat: move profile endpoints to Profiles API
