+++ armadillo/src/main/java/org/molgenis/armadillo/controller/AdminController.java
@@ -292,95 +292,4 @@ public class AdminController {
   public void userDelete(Principal principal, @PathVariable String email) {
     auditor.audit(() -> metadata.userDelete(email), principal, DELETE_USER, Map.of(EMAIL, email));
   }
-
-  @Operation(summary = "List profiles")
-  @ApiResponses(
-      value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "All profiles listed",
-            content =
-                @Content(
-                    array = @ArraySchema(schema = @Schema(implementation = ProfileConfig.class)))),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true)))
-      })
-  @GetMapping(path = "profiles", produces = APPLICATION_JSON_VALUE)
-  @ResponseStatus(OK)
-  public List<ProfileConfig> profileList(Principal principal) {
-    return auditor.audit(
-        () ->
-            metadata.profileList().stream()
-                .map(profile -> metadata.profileByName(profile.getName()))
-                .toList(),
-        principal,
-        LIST_PROFILES,
-        Map.of());
-  }
-
-  @Operation(summary = "Get profile by name")
-  @ApiResponses(
-      value = {
-        @ApiResponse(
-            responseCode = "200",
-            description = "Profile listed",
-            content = @Content(schema = @Schema(implementation = ProfileConfig.class))),
-        @ApiResponse(
-            responseCode = "404",
-            description = "Profile does not exist",
-            content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true)))
-      })
-  @GetMapping(value = "profiles/{name}", produces = APPLICATION_JSON_VALUE)
-  @ResponseStatus(OK)
-  public ProfileConfig profileGetByProfileName(Principal principal, @PathVariable String name) {
-    return auditor.audit(
-        () -> metadata.profileByName(name), principal, GET_PROFILE, Map.of(PROFILE, name));
-  }
-
-  @Operation(summary = "Add or update profile (if enabled, including docker image)")
-  @ApiResponses(
-      value = {
-        @ApiResponse(responseCode = "204", description = "Profile added or updated"),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true)))
-      })
-  @PutMapping(value = "profiles", produces = TEXT_PLAIN_VALUE)
-  @ResponseStatus(OK)
-  public void profileUpsert(Principal principal, @RequestBody ProfileConfig profileConfig)
-      throws InterruptedException {
-    metadata.profileUpsert(profileConfig);
-    auditor.audit(
-        () -> {}, // because of the exception that might happen, and cannot be caught
-        principal,
-        UPSERT_PROFILE,
-        Map.of(PROFILE, profileConfig));
-  }
-
-  @Operation(summary = "Delete profile (if enabled, including docker image)")
-  @ApiResponses(
-      value = {
-        @ApiResponse(responseCode = "204", description = "Profile deleted"),
-        @ApiResponse(
-            responseCode = "404",
-            description = "Profile does not exist",
-            content = @Content(schema = @Schema(hidden = true))),
-        @ApiResponse(
-            responseCode = "401",
-            description = "Unauthorized",
-            content = @Content(schema = @Schema(hidden = true)))
-      })
-  @DeleteMapping(value = "profiles/{name}", produces = TEXT_PLAIN_VALUE)
-  @ResponseStatus(NO_CONTENT)
-  public void profileDelete(Principal principal, @PathVariable String name) {
-    auditor.audit(
-        () -> metadata.profileDelete(name), principal, DELETE_PROFILE, Map.of(PROFILE, name));
-  }
 }
